Added more commands, starting to do more logic in connection service, and added more error handling to queryservice. Also made queryresult serializable which will be needed for the connectionservice.

+++ NoiseDB/NoiseDB/QueryService.cs
@@ -10,6 +10,7 @@ namespace NoiseDB
     {
         private IDataService DataService;
         private ConnectionService ConnectionService;
+        private bool IsConnectedToNetworkDataStore = false;
 
         public QueryService(IDataService dataService, ConnectionService connectionService)
         {
@@ -41,14 +42,54 @@ namespace NoiseDB
             switch(query.Command)
             {
                 case Commands.GET:
+                    if (string.IsNullOrEmpty(query.Key))
+                    {
+                        return new QueryResult("Failed", new ArgumentNullException(), null);
+                    }
                     return DataService.GetRow(query.Key);
+
                 case Commands.SET:
-                    return DataService.SetValue(query.Key, query.Argument);                    
+                    if (string.IsNullOrEmpty(query.Argument) || string.IsNullOrEmpty(query.Key))
+                    {
+                        return new QueryResult("Failed", new ArgumentNullException(), null);
+                    }
+                    return DataService.SetValue(query.Key, query.Argument);  
+                  
                 case Commands.DELETE:
+                    if (string.IsNullOrEmpty(query.Key))
+                    {
+                        return new QueryResult("Failed", new ArgumentNullException(), null);
+                    }
                     return DataService.DeleteRow(query.Key);
-                case Commands.SERVER_START:
+
+                case Commands.SERVER_START:                    
                     return ConnectionService.ListenForConnection();
+
                 case Commands.SERVER_STOP:
+                    return new QueryResult("NotImplemented", null, null);      
+              
+                case Commands.SERVER_CONNECT:
+                    IsConnectedToNetworkDataStore = true;
+                    return new QueryResult("NotImplemented", null, null);
+
+                case Commands.SERVER_DISCONNECT:
+                    IsConnectedToNetworkDataStore = false;
+                    return new QueryResult("NotImplemented", null, null);
+
+                case Commands.SAVE:
+                    if (string.IsNullOrEmpty(query.Key))
+                    {
+                        return new QueryResult("Failed", new ArgumentNullException(), null);
+                    }                    
+                    return DataService.SaveStore(query.Key);     
+               
+                case Commands.LOAD:
+                    if (string.IsNullOrEmpty(query.Key))
+                    {
+                        return new QueryResult("Failed", new ArgumentNullException(), null);
+                    }                    
+                    return DataService.LoadStore(query.Key);
+
                 default:
                     return new QueryResult("Unrecognised command", null, null);                
             }
